Updated documentation around helpers
I was trying to add helpers to a pug and I couldn't get it to work because the helper that I added kept saying it was `undefined`. So I tried the same code with a different language like handlebars and it worked as expected.

Then when I was poking around a little bit more I found that there was a variable that was created called `helpers` and in that object were the helpers I was trying to add.